Refine user agent context
