--- packages/triple-web/src/contexts/user-agent.ts
@@ -1,17 +1,0 @@
-import { createContext, useContext } from 'react'
-
-import { UserAgentValue } from '../types'
-
-export const UserAgentContext = createContext<UserAgentValue | undefined>(
-  undefined,
-)
-
-export function useUserAgent() {
-  const context = useContext(UserAgentContext)
-
-  if (context === undefined) {
-    throw new Error('UserAgentContext가 없습니다.')
-  }
-
-  return context
-}
