+++ packages/triple-web/src/hooks/user-agent/use-user-agent.ts
@@ -1,1 +1,13 @@
-export { useUserAgent } from '../../contexts'
+import { useContext } from 'react'
+
+import { UserAgentContext } from '../../contexts/user-agent'
+
+export function useUserAgent() {
+  const context = useContext(UserAgentContext)
+
+  if (context === undefined) {
+    throw new Error('UserAgentContext가 없습니다.')
+  }
+
+  return context
+}
